some docs, tests and cleaning up
