+++ kotlintest-assertions-arrow/src/main/kotlin/io/kotlintest/assertions/arrow/gen/extensions.kt
@@ -2,7 +2,6 @@ package io.kotlintest.assertions.arrow.gen
 
 import arrow.Kind
 import arrow.core.Either
-import arrow.core.fix
 import arrow.extension
 import arrow.typeclasses.*
 import io.kotlintest.properties.ForGen
@@ -10,6 +9,14 @@ import io.kotlintest.properties.Gen
 import io.kotlintest.properties.GenOf
 import io.kotlintest.properties.fix
 
+/**
+ * [Semigroup] extension for Gen<A>
+ *
+ * Combine two generators of type Gen<A> into a single Gen<A> provided
+ * there is a evidence of a Semigroup<A>
+ *
+ * (Gen<A>, Gen<A>, with Semigroup<A>) -> Gen<A>
+ */
 @extension
 interface GenSemigroup<A> : Semigroup<Gen<A>> {
   fun SA(): Semigroup<A>
@@ -24,6 +31,15 @@ interface GenSemigroup<A> : Semigroup<Gen<A>> {
     }
 }
 
+/**
+ * [Monoid] extension for Gen<A>
+ *
+ * Provide an [empty] [Gen] for [A] provided there is an existing
+ * [Monoid] for [A]
+ *
+ * (Gen<A>, Gen<A>, with Semigroup<A>) -> Gen<A>
+ * (with Monoid<A>) -> Gen<A>
+ */
 @extension
 interface GenMonoid<A> : Monoid<Gen<A>>, GenSemigroup<A> {
   fun MA(): Monoid<A>
@@ -33,14 +49,32 @@ interface GenMonoid<A> : Monoid<Gen<A>>, GenSemigroup<A> {
   override fun empty(): Gen<A> = Gen.constant(MA().empty())
 }
 
+/**
+ * [Functor] extension for Gen<_>
+ *
+ * Transform a `Gen<A>` into a `Gen<B>` given a `(A) -> B` function
+ *
+ * (Gen<A>, (A) -> B) -> Gen<B>
+ */
 @extension
 interface GenFunctor : Functor<ForGen> {
   override fun <A, B> GenOf<A>.map(f: (A) -> B): Gen<B> =
     fix().map(f)
 }
 
+/**
+ * [Applicative] extension for Gen<_>
+ *
+ * Lift a value [A] in the environment into a Gen<A> via [just]
+ *
+ * (A) -> Gen<A>
+ *
+ * Map multiple [Gen] with arbitrary arity to compose new Generators with [map]
+ *
+ * (Gen<A>, Gen<B>, Gen<C>, ...) { (a, b, c, d) -> ... Gen<Z> }
+ */
 @extension
-interface GenApplicative : Applicative<ForGen> {
+interface GenApplicative : Applicative<ForGen>, GenFunctor {
   override fun <A, B> GenOf<A>.ap(ff: Kind<ForGen, (A) -> B>): Gen<B> =
     ff.fix().flatMap { this.fix().map(it) }
 
@@ -51,12 +85,28 @@ interface GenApplicative : Applicative<ForGen> {
     Gen.constant(a)
 }
 
+/**
+ * [Monad] extension for Gen<_>
+ *
+ * Sequentially bind generators to produce new generators
+ *
+ * ```kotlin
+ * val prefix = "_"
+ * val personGen: Gen<Person> =
+ *   binding {
+ *     val id = Gen.long().bind()
+ *     val name = Gen.string().bind()
+ *     Person(id, prefix + name)
+ *   }
+ * forAll(personGen) { it.name.startsWith(prefix) }
+ * ```
+ */
 @extension
-interface GenMonad : Monad<ForGen> {
+interface GenMonad : Monad<ForGen>, GenApplicative {
   override fun <A, B> GenOf<A>.ap(ff: Kind<ForGen, (A) -> B>): Gen<B> =
     fix().ap(ff)
 
-  override fun <A, B> GenOf<A>.flatMap(f: (A) -> Kind<ForGen, B>): Gen<B> =
+  override fun <A, B> GenOf<A>.flatMap(f: (A) -> GenOf<B>): Gen<B> =
     fix().flatMap(f)
